refactor(ui-kit): remove cycle dependencies
