+++ .changeset/spotty-chicken-film.md
@@ -1,0 +1,5 @@
+---
+'@iota/apps-ui-kit': patch
+---
+
+Update imports to avoid cycle dependencies
