+++ apps/ui-kit/src/lib/components/molecules/navbar-item/NavbarItemHorizontal.tsx
@@ -13,7 +13,7 @@ import {
     UNSELECTED_ICON,
     UNSELECTED_TEXT,
 } from './navbarItem.classes';
-import { Badge, BadgeType } from '../../atoms';
+import { Badge, BadgeType } from '@/components/atoms';
 import type { NavbarItemProps } from './NavbarItem';
 
 export function NavbarItemHorizontal({
