+++ apps/ui-kit/src/lib/components/organisms/table/Table.tsx
@@ -4,8 +4,9 @@
 import type { PropsWithChildren, ReactNode } from 'react';
 import cx from 'classnames';
 import { TableProvider, useTableContext } from './TableContext';
-import type { ButtonProps } from '@/lib';
-import { Button, ButtonSize, ButtonType, Checkbox, TableCellBase, TableHeaderCell } from '@/lib';
+import type { ButtonProps } from '@/components/atoms';
+import { Button, ButtonSize, ButtonType, Checkbox } from '@/components/atoms';
+import { TableCellBase, TableHeaderCell } from '@/components/molecules';
 import { ArrowLeft, DoubleArrowLeft, ArrowRight, DoubleArrowRight } from '@iota/apps-ui-icons';
 
 export interface TablePaginationOptions {
